+++ tests/Phan/PhanTest7.php
@@ -1,0 +1,14 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Phan\Tests;
+
+/**
+ * Runs tests/files/src/0700-0799
+ */
+class PhanTest7 extends PhanTestRange
+{
+    public const START_RANGE = '0700';
+    public const END_RANGE   = '0800';
+}